s/command/type in webview and other refactors
